Upload images (#2)
* update schema

* get upload vs url working in form

* use cva in image component

* add image source switching logic to item image component

* small tweaks

* tweaks

* add file to item insert

* create image dropzone

* add image dropzone hook

* create item image droppable

* write upload code

* get image upload working!!!

* remove item image dialog

* update init functions

* add image input options

* update item form image input

* ensure drag and drop works

* allow for deleting of images

* update item insert

* get items updating

* get create working

* upload image upon item creation

* add formdata utilities

* create separate action for image handling

* update mutations

* fix type errors

* fix type errors

* get update and create working

+++ src/actions/items/items.handlers.ts
@@ -1,5 +1,5 @@
 import { Category, CategoryItem, Item, List } from "@/db/schema";
-import  { createDb } from "@/db";
+import { createDb } from "@/db";
 import { and, eq } from "drizzle-orm";
 
 import { idAndUserIdFilter } from "@/actions/filters";
@@ -21,11 +21,12 @@ const getAll: ActionHandler<typeof itemInputs.getAll, ItemSelect[]> = async (
 };
 
 const create: ActionHandler<typeof itemInputs.create, ItemSelect> = async (
-  { data },
+  data,
   c,
 ) => {
   const db = createDb(c.locals.runtime.env);
   const userId = isAuthorized(c).id;
+
   const [newItem] = await db
     .insert(Item)
     .values({ ...data, userId, id: uuid() })
@@ -65,23 +66,54 @@ const remove: ActionHandler<typeof itemInputs.remove, null> = async (
 ) => {
   const db = createDb(c.locals.runtime.env);
   const userId = isAuthorized(c).id;
+
+  const [item] = await db
+    .select()
+    .from(Item)
+    .where(idAndUserIdFilter(Item, { userId, id: itemId }));
+
+  if (!item) {
+    throw new ActionError({
+      code: "NOT_FOUND",
+      message: "Item not found",
+    });
+  }
+
+  if (item.imageR2Key) {
+    await c.locals.runtime.env.R2_BUCKET.delete(item.imageR2Key);
+  }
+
   await db.delete(CategoryItem).where(eq(CategoryItem.itemId, itemId));
   await db.delete(Item).where(idAndUserIdFilter(Item, { userId, id: itemId }));
   return null;
 };
 
 const update: ActionHandler<typeof itemInputs.update, ItemSelect> = async (
-  { itemId, data },
+  data,
   c,
 ) => {
   const db = createDb(c.locals.runtime.env);
   const userId = isAuthorized(c).id;
-  const updated = await db
+
+  const { id: itemId } = data;
+
+  const [item] = await db
+    .select()
+    .from(Item)
+    .where(idAndUserIdFilter(Item, { userId, id: itemId }));
+
+  if (!item) {
+    throw new ActionError({
+      code: "NOT_FOUND",
+      message: "Item not found",
+    });
+  }
+
+  const [updated] = await db
     .update(Item)
     .set(data)
     .where(idAndUserIdFilter(Item, { userId, id: itemId }))
-    .returning()
-    .then((rows) => rows[0]);
+    .returning();
   return updated;
 };
 
@@ -104,6 +136,51 @@ const getListsIncluded: ActionHandler<
   return result;
 };
 
+const imageUpload: ActionHandler<typeof itemInputs.imageUpload, null> = async (
+  { itemId, imageFile, removeImageFile },
+  c,
+) => {
+  const db = createDb(c.locals.runtime.env);
+  const userId = isAuthorized(c).id;
+
+  const [item] = await db
+    .select({ id: Item.id, userId: Item.userId, imageKey: Item.imageR2Key })
+    .from(Item)
+    .where(idAndUserIdFilter(Item, { userId, id: itemId }));
+
+  if (!item) {
+    throw new ActionError({
+      code: "NOT_FOUND",
+      message: "Item not found",
+    });
+  }
+
+  if (item.userId !== userId) {
+    throw new ActionError({
+      code: "UNAUTHORIZED",
+      message: "You are not authorized to update this item",
+    });
+  }
+
+  if (imageFile && imageFile.size > 0) {
+    // TODO: downsize and convert image
+
+    const key = crypto.randomUUID();
+    await c.locals.runtime.env.R2_BUCKET.put(key, imageFile);
+    await db
+      .update(Item)
+      .set({ imageR2Key: key, imageType: "file" })
+      .where(eq(Item.id, itemId));
+  }
+
+  if (removeImageFile && item.imageKey) {
+    await c.locals.runtime.env.R2_BUCKET.delete(item.imageKey);
+    await db.update(Item).set({ imageR2Key: null }).where(eq(Item.id, itemId));
+  }
+
+  return null;
+};
+
 const itemHandlers = {
   getAll,
   create,
@@ -111,6 +188,7 @@ const itemHandlers = {
   remove,
   update,
   getListsIncluded,
+  imageUpload,
 };
 
 export default itemHandlers;
